fix: change float to decimal in location class, add more validation

+++ src/main/java/com/ptit/thesis/smartrecruit/repository/LocationRepository.java
@@ -1,6 +1,7 @@
 package com.ptit.thesis.smartrecruit.repository;
 
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,6 +11,6 @@ import com.ptit.thesis.smartrecruit.entity.Location;
 
 @Repository
 public interface LocationRepository extends JpaRepository<Location, Long> {
-    Optional<Location> findByLatitudeAndLongitude(Float latitude, Float longitude);
-    boolean existsByLatitudeAndLongitude(Float latitude, Float longitude);
+    Optional<Location> findByLatitudeAndLongitude(BigDecimal latitude, BigDecimal longitude);
+    boolean existsByLatitudeAndLongitude(BigDecimal latitude, BigDecimal longitude);
 }
